stock_app/requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
Django==5.2.5
django-formtools==2.5.1
django-phonenumber-field==8.2.0
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
mysqlclient==2.2.7
openpyxl==3.1.5
phonenumbers==9.0.15
pillow==11.3.0
reportlab==4.4.4
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
gunicorn==23.0.0
